reagent: adapt codebase structure & cleanup unused classes
