+++ inseq/attr/feat/ops/reagent_core/stopping_condition_evaluator.py
@@ -1,12 +1,33 @@
 import logging
+from abc import ABC, abstractmethod
 
 import torch
 from transformers import AutoModelWithLMHead, AutoTokenizer
 from typing_extensions import override
 
-from ..token_replacement.token_replacer.ranking import RankingTokenReplacer
-from ..token_replacement.token_sampler.base import TokenSampler
-from .base import StoppingConditionEvaluator
+from .token_replacer import RankingTokenReplacer
+from .token_sampler import TokenSampler
+
+
+class StoppingConditionEvaluator(ABC):
+    """Base class for Stopping Condition Evaluators"""
+
+    @abstractmethod
+    def __call__(
+        self, input_ids: torch.Tensor, target_id: torch.Tensor, importance_score: torch.Tensor
+    ) -> torch.Tensor:
+        """Evaluate stop condition according to the specified strategy.
+
+        Args:
+            input_ids: Input sequence [batch, sequence]
+            target_id: Target token [batch]
+            importance_score: Importance score of the input [batch, sequence]
+
+        Return:
+            Whether the stop condition achieved [batch]
+
+        """
+        raise NotImplementedError()
 
 
 class TopKStoppingConditionEvaluator(StoppingConditionEvaluator):
@@ -37,8 +58,6 @@ class TopKStoppingConditionEvaluator(StoppingConditionEvaluator):
             tokenizer: (Optional) Used for logging top_k_words at each step
 
         """
-        super().__init__()
-
         self.model = model
         self.token_sampler = token_sampler
         self.top_k = top_k
@@ -46,7 +65,7 @@ class TopKStoppingConditionEvaluator(StoppingConditionEvaluator):
         self.tokenizer = tokenizer
 
     @override
-    def evaluate(
+    def __call__(
         self, input_ids: torch.Tensor, target_id: torch.Tensor, importance_score: torch.Tensor
     ) -> torch.Tensor:
         """Evaluate stop condition
@@ -60,12 +79,10 @@ class TopKStoppingConditionEvaluator(StoppingConditionEvaluator):
             Whether the stop condition achieved [batch]
 
         """
-        super().evaluate(input_ids, target_id, importance_score)
-
         # Replace tokens with low importance score and then inference \hat{y^{(e)}_{t+1}}
 
         self.token_replacer.set_score(importance_score)
-        input_ids_replaced, mask_replacing = self.token_replacer.sample(input_ids)
+        input_ids_replaced, mask_replacing = self.token_replacer(input_ids)
 
         logging.debug(f"Replacing mask based on importance score -> { mask_replacing }")
 
@@ -74,9 +91,6 @@ class TopKStoppingConditionEvaluator(StoppingConditionEvaluator):
         assert not input_ids_replaced.requires_grad, "Error: auto-diff engine not disabled"
         with torch.no_grad():
             logits_replaced = self.model(input_ids_replaced)["logits"]
-
-        if self.trace_target_likelihood is not None:
-            self.trace_target_likelihood.append(torch.softmax(logits_replaced, dim=-1)[:, -1, target_id])
 
         ids_prediction_sorted = torch.argsort(logits_replaced[:, -1, :], descending=True)
         ids_prediction_top_k = ids_prediction_sorted[:, : self.top_k]
@@ -91,18 +105,33 @@ class TopKStoppingConditionEvaluator(StoppingConditionEvaluator):
         # Stop flags for each sample in the batch
         return match_hit
 
+
+class DummyStoppingConditionEvaluator(StoppingConditionEvaluator):
+    """
+    Stopping Condition Evaluator which stop when target exist in top k predictions,
+    while top n tokens based on importance_score are not been replaced.
+    """
+
     @override
-    def trace_start(self) -> None:
-        """Start tracing"""
-        super().trace_start()
-
-        self.token_sampler.trace_start()
-        self.token_replacer.trace_start()
+    def __init__(self) -> None:
+        """Constructor"""
 
     @override
-    def trace_stop(self) -> None:
-        """Stop tracing"""
-        super().trace_stop()
+    def __call__(
+        self, input_ids: torch.Tensor, target_id: torch.Tensor, importance_score: torch.Tensor
+    ) -> torch.Tensor:
+        """Evaluate stop condition
 
-        self.token_sampler.trace_stop()
-        self.token_replacer.trace_stop()
+        Args:
+            input_ids: Input sequence [batch, sequence]
+            target_id: Target token [batch]
+            importance_score: Importance score of the input [batch, sequence]
+
+        Return:
+            Whether the stop condition achieved [batch]
+
+        """
+        match_hit = torch.ones([input_ids.shape[0]], dtype=torch.bool, device=input_ids.device)
+
+        # Stop flags for each sample in the batch
+        return match_hit
