--- inseq/attr/feat/ops/reagent_core/utils/traceable.py
@@ -1,8 +1,0 @@
-class Traceable:
-    """Traceable base"""
-
-    def trace_start(self) -> None:
-        """Base trace_start"""
-
-    def trace_stop(self) -> None:
-        """Base trace_stop"""
